[sui] Remove sui -> sui-adapter dependency
It was only used to set a test-only override, which has now been moved
to a module in `sui-types`.  This makes one fewer crate that depends
directly on `sui-adapter` (and one fewer crate that needs to handle
its versioned API).

Test Plan:

```
crates/sui$ cargo simtest
```
