+++ crates/sui/Cargo.toml
@@ -28,7 +28,6 @@ rand = "0.8.5"
 tap = "1.0"
 inquire = "0.6.0"
 
-sui-adapter = { path = "../sui-adapter" }
 sui-core = { path = "../sui-core" }
 sui-framework-snapshot = { path = "../sui-framework-snapshot" }
 sui-config = { path = "../sui-config" }
